cfg-python: Use GenericCFGBuilder
